* commit 'd1b9241a91f5eef9995e54cf80f9789045e5a233':
  Object model changes to support 64bit.

+++ compiler/output_stream.h
@@ -41,7 +41,7 @@ class OutputStream {
     return location_;
   }
 
-  virtual bool WriteFully(const void* buffer, int64_t byte_count) = 0;
+  virtual bool WriteFully(const void* buffer, size_t byte_count) = 0;
 
   virtual off_t Seek(off_t offset, Whence whence) = 0;
